pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xocto"
version = "4.9.0"
requires-python = ">=3.9"
description = "Kraken Technologies Python service utilities"
readme = "README.md"
authors = [
  {name = "Kraken Technologies", email = "talent@octopus.energy"},
]
maintainers = [
  {name = "Kraken Technologies", email = "talent@octopus.energy"},
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
license = {text = "MIT"}
dependencies = [
  "ddtrace>=1.9.0",
  "duckdb>=0.9.0",
  "django>=4.0",
  "openpyxl>=3.1.0",
  "pact-python>=1.6.0",
  "pandas>=1.5.3",
  "pyarrow>=11.0.0",
  "python-dateutil>=2.8.2",
  "python-magic>=0.4.27",
  "pytz",
  "structlog>=20.2.0",
  "xlrd>=2.0.1",
]

[project.optional-dependencies]
dev = [
  "boto3==1.26.53",
  "botocore==1.29.53",
  "mypy-boto3-s3==1.26.0.post1",
  "mypy==0.991",
  "numpy==1.22.2",
  "pre-commit>=3.2.0",
  "pyarrow-stubs==10.0.1.6",
  "ruff==0.1.3",
  "twine==4.0.2",
  "types-openpyxl==3.0.4.5",
  "types-python-dateutil==2.8.19.6",
  "types-pytz==2022.7.1.0",
  "types-requests==2.28.11.8",
  "wheel==0.38.4",
]
test = [
  "hypothesis==6.62.1",
  "moto[s3,sqs]==4.1",
  "ruff==0.1.3",  
  "pytest-django==4.5.2",
  "pytest-mock==3.10.0",
  "pytest==7.2.1",
  "time-machine==2.9.0",
]
docs = [
  "Sphinx==4.5.0",
  "myst-parser==0.18.1",
]

[project.urls]
changelog = "https://github.com/octoenergy/xocto/blob/main/CHANGELOG.md"
documentation = "https://xocto.readthedocs.io"
issues = "https://github.com/octoenergy/xocto/issues"

[tool.setuptools]
packages = ["xocto", "xocto.events", "xocto.storage"]
# Old setup.py version required
#  package_data={"xocto": ["py.typed"]},

[tool.mypy]
# Specify which files to check.
files = [
  "xocto",
  "tests",
  "*.py",
]

# Output.
show_error_codes = true

# Strict mode; enables the following flags: 
#   --check-untyped-defs
#   --disallow-any-generics,
#   --disallow-incomplete-defs
#   --disallow-subclassing-any 
#   --disallow-untyped-calls 
#   --disallow-untyped-decorators,
#   --disallow-untyped-defs,
#   --no-implicit-reexport
#   --strict-concatenate
#   --strict-equality
#   --warn-redundant-casts
#   --warn-return-any
#   --warn-unused-configs 
#   --warn-unused-ignores
strict = true

# These flags enabled by `strict` are ones we're overriding.
disallow_subclassing_any = false
warn_return_any = false

# Enable the following optional checks.
enable_error_code = [
  "ignore-without-code",
  "redundant-expr",
  "truthy-bool",
  "unused-awaitable",
]

# Warnings.
warn_unreachable = true

# Don't force test functions to include types.
[[tool.mypy.overrides]]
module = [
  "tests.*",
]
disallow_untyped_defs = false
check_untyped_defs = false

# Ignore errors from packages that don't have type annotations.
[[tool.mypy.overrides]]
module = [
  "boto3.*",
  "botocore.*",
  "django.*",
  "moto.*",
  "pact.*",
  "pandas.*",
  "setuptools.*",
  "xlrd.*",
  "zoneinfo.*",
]
ignore_missing_imports = true


# Ruff
# ----

[tool.ruff]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501",  # line too long - black takes care of this for us
]

[tool.ruff.per-file-ignores]
# Allow unused imports in __init__ files as these are convenience imports
"**/__init__.py" = [ "F401" ]

[tool.ruff.isort]
lines-after-imports = 2
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "project",
    "local-folder",
]

[tool.ruff.isort.sections]
"project" = [
    "xocto",
    "tests",
]
